Add overall styling to the app
Add styling to to all of the components
