+++ package.json
@@ -20,6 +20,8 @@
   },
   "homepage": "https://github.com/ermin-cahtarevic/weather-app#readme",
   "devDependencies": {
+    "css-loader": "^3.5.2",
+    "style-loader": "^1.1.3",
     "webpack": "^4.42.1",
     "webpack-cli": "^3.3.11"
   }
