+++ src/css/style.css
@@ -1,0 +1,134 @@
+body {
+  background-color: rgb(19, 37, 71);
+  background-image: url('https://images.unsplash.com/photo-1526451752866-695bdb5512ca?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1051&q=80');
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center center;
+  margin: 0;
+  padding: 0;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-family: Arial, Helvetica, sans-serif;
+}
+
+#main {
+  border: 4px solid white;
+  border-radius: 10px;
+  min-width: 900px;
+  max-width: 900px;
+  height: 700px;
+  max-height: 90vh;
+  background-color: rgba(19, 37, 71, 0.7);
+}
+
+.search-wrap {
+  height: 200px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.search-form {
+  margin-top: 35px;
+}
+
+.search-form input {
+  font-size: 16px;
+  margin: 0;
+  padding: 7px 15px;
+  background-color: rgb(19, 37, 71);
+  color: white;
+  outline: none;
+  border: 2px solid white;
+  border-radius: 20px;
+}
+
+.search-form input:focus {
+  background-color: white;
+  color: rgb(10, 18, 34);
+}
+
+.btn {
+  background-color: rgb(19, 37, 71);
+  color: white;
+  outline: none;
+  border: 2px solid white;
+  border-radius: 20px;
+  font-size: 16px;
+  padding: 7px 15px;
+  margin: 0;
+  transition: all 0.3s ease-out;
+}
+
+.btn:hover {
+  background-color: white;
+  color: rgb(19, 37, 71);
+  cursor: pointer;
+}
+
+.data-wrap {
+  color: white;
+  width: 100%;
+  height: 500px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding-bottom: 50px;
+}
+
+.location {
+  font-size: 24px;
+  margin-bottom: 60px;
+}
+
+.temperature-wrap {
+  width: 100%;
+  text-align: center;
+}
+
+.temperature {
+  font-weight: bold;
+}
+
+.temperature-wrap .temperature,
+.temperature-wrap .unit-toggle {
+  font-size: 55px;
+}
+
+.temperature-wrap .unit-toggle {
+  text-decoration: none;
+  color: white;
+}
+
+.weather-type {
+  margin-top: 15px;
+  font-size: 22px;
+}
+
+.misc-wrap {
+  display: flex;
+  justify-content: space-between;
+  width: 500px;
+  margin-top: 80px;
+}
+
+.humidity-wrap,
+.wind-wrap {
+  text-align: center;
+}
+
+.humidity-wrap i,
+.wind-wrap i {
+  margin-right: 5px;
+}
+
+.humidity-wrap i,
+.wind-wrap i,
+.humidity-wrap .humidity,
+.wind-wrap .wind {
+  font-size: 28px;
+}
+
